+++ tests/render/app/route.js
@@ -7,13 +7,27 @@ Router.setView( "engine", handlebars );
 Router.setView( "view", path.join( __dirname, "views" ));
 Router.setView( "layout", path.join( __dirname, "views", "layout" ) );
 Router.setView( "ext", ".hbs" );
+Router.setView( "default", [
+  {
+    name: "layout",
+    format: "static",
+    replaceable: false,
+    content: "main"
+  },
+  {
+    name: "partials/navbar",
+    format: "dynamic",
+    replaceable: true,
+    parser: () => path.join( "partials", "navbar" )
+  }
+]);
 
 Router.get("/", (req, res) => {
   res.render("index");
 });
 
 Router.get("/layout", ( req, res ) => {
-  res.render( "index", { layout: "main" } );
+  res.render( "index", { layout: "second" } );
 })
 
 module.exports = Router;